[TimeManager]: Make TimeManager use milliseconds
The TimeManager used by Anvil uses seconds for internal time keeping.
We want to reuse as much code as possible in order to keep the code
base small and easy to maintain. But we can not reuse the TimeManager
for anvil-polkadot as the pallet_timestamp expect monotonically increasing
milliseconds.
Make the TimeManager use milliseconds internally and return the
`next_timestamp` as either seconds or milliseconds depending on
the desire precision.

Signed-off-by: Alexandru Cihodaru <alexandru.cihodaru@parity.io>