Don't derenference SubstrateBlock on revert
update_latest took ownership of the Substrateblock and wrapped it
in Arc::new. During the rollback/revert operations a caller might
try to update the latest block to a block obtained by block_by_number.
If one of the "Best" or "Finalized" subscription is lagging this could
make block_by_number to return an strong reference. This prevents
the caller to update the best block since it will not be able to
unwrap the strong reference.

Signed-off-by: Alexandru Cihodaru <alexandru.cihodaru@parity.io>